Address outstanding User Comments Client TODOs
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/WDKClient/trunk@24537 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ packages/libs/wdk-client/src/Controllers/UserCommentShowController.tsx
@@ -297,7 +297,7 @@ const mergeProps = (
                           ({ start, end }) => `${start}-${end}`
                         ).join(', ')
                       }
-                      {comment.location.reversed && ` (reversed)`}
+                      {comment.location.reverse && ` (reversed)`}
                     </>
                   )
                   : null
@@ -415,7 +415,9 @@ const mergeProps = (
                   {' '}
                   <Link to={`/user-comments/delete?commentId=${comment.id}`} onClick={(event: React.MouseEvent<HTMLAnchorElement>) => {
                     event.preventDefault();
-                    deleteUserComment(comment.id);
+                    if (confirm(`Are you sure you wish to delete comment ${comment.id}?`)) {
+                      deleteUserComment(comment.id);
+                    }
                   }}>
                     [delete comment]
                   </Link>
